Attempt to move private classes to a dedicated package
